refactor: extract SimilarIssueMatch.swift from IssueExportReview.swift
Splits IssueExportReview.swift by concern:
- IssueExportReview.swift keeps the review workspace types
  (IssueExportReviewItem + IssueExportReview)
- SimilarIssueMatch.swift gets the similarity types
  (SimilarIssueResolution enum + SimilarIssueMatch struct)

Byte-identical move of lines 3-51. Both files import Foundation.

IssueExportReview.swift: 130 → 80 lines. Tests: 667.
